Move ResourceView to resource.rs
