+++ src/controller_examples/simple_controller/proof/safety.rs
@@ -7,7 +7,7 @@ use crate::controller_examples::simple_controller::spec::{
     reconciler,
     reconciler::{simple_reconciler, SimpleReconcileState},
 };
-use crate::kubernetes_api_objects::{api_method::*, common::*, config_map::*, dynamic::*};
+use crate::kubernetes_api_objects::{api_method::*, common::*, config_map::*, resource::*};
 use crate::kubernetes_cluster::{
     proof::{
         controller_runtime_liveness::reconciler_init_and_no_pending_req, controller_runtime_safety,
