+++ src/controller_examples/simple_controller/proof/shared.rs
@@ -6,7 +6,7 @@ use crate::controller_examples::simple_controller::spec::{
     reconciler,
     reconciler::{simple_reconciler, SimpleReconcileState},
 };
-use crate::kubernetes_api_objects::{common::*, config_map::*, dynamic::*};
+use crate::kubernetes_api_objects::{common::*, config_map::*, resource::*};
 use crate::kubernetes_cluster::spec::{
     controller::common::{controller_req_msg, ControllerAction, ControllerActionInput},
     distributed_system::*,
